Split operations within 2 different classes. General refactoring.
