+++ pom.xml
@@ -6,7 +6,7 @@
     <modelVersion>4.0.0</modelVersion>
     <groupId>com.mule.mulechain</groupId>
     <artifactId>mulechain-vectors</artifactId>
-    <version>0.1.65-SNAPSHOT</version>
+    <version>0.1.68-SNAPSHOT</version>
     <packaging>mule-extension</packaging>
     <name>MAC Vectors</name>
 
@@ -21,7 +21,8 @@
 		<formatterConfigPath>formatter.xml</formatterConfigPath>
 		<javaFormatter.plugin.version>2.0.1</javaFormatter.plugin.version>
 		<formatterGoal>validate</formatterGoal>	
-		<langchain4jVersion>0.35.0</langchain4jVersion>	
+		<langchain4jVersion>0.35.0</langchain4jVersion>
+		<slf4jApi.version>2.0.7</slf4jApi.version>
     </properties>
 
 <!-- 	<build>
@@ -96,7 +97,11 @@
 			<artifactId>langchain4j</artifactId>
 			<version>${langchain4jVersion}</version>
 		</dependency>
-
+		<dependency>
+			<groupId>org.slf4j</groupId>
+			<artifactId>slf4j-api</artifactId>
+			<version>${slf4jApi.version}</version>
+		</dependency>
 		<dependency>
 			<groupId>org.json</groupId>
 			<artifactId>json</artifactId>
