+++ src/main/java/org/mule/extension/mulechain/vectors/internal/config/Configuration.java
@@ -1,9 +1,9 @@
 package org.mule.extension.mulechain.vectors.internal.config;
 
-import org.mule.extension.mulechain.vectors.internal.helpers.providers.MuleChainVectorsEmbeddingModelServiceProvider;
-import org.mule.extension.mulechain.vectors.internal.helpers.providers.MuleChainVectorsVectorStoreProvider;
-import org.mule.extension.mulechain.vectors.internal.operation.MuleChainVectorsOperations;
-import org.mule.runtime.extension.api.annotation.Configuration;
+import org.mule.extension.mulechain.vectors.internal.helpers.providers.EmbeddingModelServiceProvider;
+import org.mule.extension.mulechain.vectors.internal.helpers.providers.VectorStoreProvider;
+import org.mule.extension.mulechain.vectors.internal.operation.DocumentOperations;
+import org.mule.extension.mulechain.vectors.internal.operation.EmbeddingOperations;
 import org.mule.runtime.extension.api.annotation.Operations;
 import org.mule.runtime.extension.api.annotation.param.Parameter;
 import org.mule.runtime.extension.api.annotation.param.display.Placement;
@@ -13,18 +13,18 @@ import org.mule.runtime.extension.api.annotation.values.OfValues;
  * This class represents an extension configuration, values set in this class are commonly used across multiple
  * operations since they represent something core from the extension.
  */
-@Configuration(name = "config")
-@Operations(MuleChainVectorsOperations.class)
-public class MuleChainVectorsConfiguration {
+@org.mule.runtime.extension.api.annotation.Configuration(name = "config")
+@Operations({EmbeddingOperations.class, DocumentOperations.class})
+public class Configuration {
 
   @Parameter
   @Placement(order = 1, tab = Placement.DEFAULT_TAB)
-  @OfValues(MuleChainVectorsEmbeddingModelServiceProvider.class)
+  @OfValues(EmbeddingModelServiceProvider.class)
   private String embeddingModelService;
 
   @Parameter
   @Placement(order = 2, tab = Placement.DEFAULT_TAB)
-  @OfValues(MuleChainVectorsVectorStoreProvider.class)
+  @OfValues(VectorStoreProvider.class)
   private String vectorStore;
 
   @Parameter
