+++ src/main/java/org/mule/extension/mulechain/vectors/internal/helpers/parameters/EmbeddingModelNameParameters.java
@@ -1,16 +1,16 @@
 package org.mule.extension.mulechain.vectors.internal.helpers.parameters;
 
-import org.mule.extension.mulechain.vectors.internal.helpers.providers.MuleChainVectorsEmbeddingModelNameProvider;
+import org.mule.extension.mulechain.vectors.internal.helpers.providers.EmbeddingModelNameProvider;
 import org.mule.runtime.api.meta.ExpressionSupport;
 import org.mule.runtime.extension.api.annotation.Expression;
 import org.mule.runtime.extension.api.annotation.param.Parameter;
 import org.mule.runtime.extension.api.annotation.values.OfValues;
 
-public class MuleChainVectorsEmbeddingModelNameParameters {
+public class EmbeddingModelNameParameters {
 
   @Parameter
   @Expression(ExpressionSupport.SUPPORTED)
-  @OfValues(MuleChainVectorsEmbeddingModelNameProvider.class)
+  @OfValues(EmbeddingModelNameProvider.class)
   private String embeddingModelName;
 
   public String getEmbeddingModelName() {
