+++ src/main/java/org/mule/extension/mulechain/vectors/internal/helpers/parameters/FileTypeParameters.java
@@ -1,6 +1,6 @@
 package org.mule.extension.mulechain.vectors.internal.helpers.parameters;
 
-import org.mule.extension.mulechain.vectors.internal.constants.MuleChainVectorsConstants;
+import org.mule.extension.mulechain.vectors.internal.constants.Constants;
 import org.mule.extension.mulechain.vectors.internal.helpers.providers.FileTypeEmbeddingProvider;
 import org.mule.runtime.api.meta.ExpressionSupport;
 import org.mule.runtime.extension.api.annotation.Expression;
@@ -13,7 +13,7 @@ public class FileTypeParameters {
   @Parameter
   @Expression(ExpressionSupport.SUPPORTED)
   @OfValues(FileTypeEmbeddingProvider.class)
-  @Optional(defaultValue = MuleChainVectorsConstants.FILE_TYPE_TEXT)
+  @Optional(defaultValue = Constants.FILE_TYPE_TEXT)
   private String fileType;
 
   public String getFileType() {
