+++ src/main/java/org/mule/extension/mulechain/vectors/internal/helpers/providers/EmbeddingModelServiceProvider.java
@@ -1,6 +1,5 @@
 package org.mule.extension.mulechain.vectors.internal.helpers.providers;
 
-
 import java.util.Set;
 
 import org.mule.extension.mulechain.vectors.internal.constants.Constants;
@@ -9,16 +8,17 @@ import org.mule.runtime.extension.api.values.ValueBuilder;
 import org.mule.runtime.extension.api.values.ValueProvider;
 import org.mule.runtime.extension.api.values.ValueResolvingException;
 
-public class FileTypeEmbeddingProvider implements ValueProvider {
+public class EmbeddingModelServiceProvider implements ValueProvider {
 
   @Override
   public Set<Value> resolve() throws ValueResolvingException {
 
     return ValueBuilder.getValuesFor(
-            Constants.FILE_TYPE_ANY,
-            Constants.FILE_TYPE_TEXT,
-            Constants.FILE_TYPE_URL,
-            Constants.FILE_TYPE_CRAWL);
-  }
+            Constants.EMBEDDING_MODEL_SERVICE_OPENAI,
+            Constants.EMBEDDING_MODEL_SERVICE_MISTRAL_AI,
+            Constants.EMBEDDING_MODEL_SERVICE_NOMIC,
+            Constants.EMBEDDING_MODEL_SERVICE_HUGGING_FACE,
+            Constants.EMBEDDING_MODEL_SERVICE_AZURE_OPENAI); //"OLLAMA", "COHERE", "AZURE_OPENAI", "HUGGING_FACE";
+}
 
 }
