+++ src/main/java/org/mule/extension/mulechain/vectors/internal/helpers/providers/MetadataFilterMethodProvider.java
@@ -1,6 +1,6 @@
 package org.mule.extension.mulechain.vectors.internal.helpers.providers;
 
-import org.mule.extension.mulechain.vectors.internal.constants.MuleChainVectorsConstants;
+import org.mule.extension.mulechain.vectors.internal.constants.Constants;
 import org.mule.runtime.api.value.Value;
 import org.mule.runtime.extension.api.values.ValueBuilder;
 import org.mule.runtime.extension.api.values.ValueProvider;
@@ -13,7 +13,7 @@ import java.util.Set;
  * Implements {@link ValueProvider} to supply a predefined set of values that can be used
  * as options for filtering operations, such as "isEqualTo" and "isNotEqualTo".
  */
-public class MuleChainVectorsMetadataFilterMethodProvider implements ValueProvider {
+public class MetadataFilterMethodProvider implements ValueProvider {
 
   /**
    * Resolves and returns a set of filter methods for metadata filtering.
@@ -26,8 +26,8 @@ public class MuleChainVectorsMetadataFilterMethodProvider implements ValueProvid
   public Set<Value> resolve() throws ValueResolvingException {
 
     return ValueBuilder.getValuesFor(
-            MuleChainVectorsConstants.METADATA_FILTER_METHOD_IS_EQUAL_TO,
-            MuleChainVectorsConstants.METADATA_FILTER_METHOD_IS_NOT_EQUAL_TO); // Additional methods can be added as needed
+            Constants.METADATA_FILTER_METHOD_IS_EQUAL_TO,
+            Constants.METADATA_FILTER_METHOD_IS_NOT_EQUAL_TO); // Additional methods can be added as needed
   }
 
 }
