+++ src/main/java/org/mule/extension/mulechain/vectors/internal/helpers/providers/MetadataKeyProvider.java
@@ -1,24 +1,24 @@
 package org.mule.extension.mulechain.vectors.internal.helpers.providers;
 
-
-import java.util.Set;
-
 import org.mule.extension.mulechain.vectors.internal.constants.Constants;
 import org.mule.runtime.api.value.Value;
 import org.mule.runtime.extension.api.values.ValueBuilder;
 import org.mule.runtime.extension.api.values.ValueProvider;
 import org.mule.runtime.extension.api.values.ValueResolvingException;
 
-public class FileTypeEmbeddingProvider implements ValueProvider {
+import java.util.Set;
+
+public class MetadataKeyProvider implements ValueProvider {
 
   @Override
   public Set<Value> resolve() throws ValueResolvingException {
 
     return ValueBuilder.getValuesFor(
-            Constants.FILE_TYPE_ANY,
-            Constants.FILE_TYPE_TEXT,
-            Constants.FILE_TYPE_URL,
-            Constants.FILE_TYPE_CRAWL);
+            Constants.METADATA_KEY_FILE_NAME,
+            Constants.METADATA_KEY_FILE_TYPE,
+            Constants.METADATA_KEY_URL,
+            Constants.METADATA_KEY_FULL_PATH,
+            Constants.METADATA_KEY_ABSOLUTE_DIRECTORY_PATH);// "textSegment"
   }
 
 }
