+++ src/main/java/org/mule/extension/mulechain/vectors/internal/helpers/providers/VectorStoreProvider.java
@@ -1,6 +1,5 @@
 package org.mule.extension.mulechain.vectors.internal.helpers.providers;
 
-
 import java.util.Set;
 
 import org.mule.extension.mulechain.vectors.internal.constants.Constants;
@@ -9,16 +8,20 @@ import org.mule.runtime.extension.api.values.ValueBuilder;
 import org.mule.runtime.extension.api.values.ValueProvider;
 import org.mule.runtime.extension.api.values.ValueResolvingException;
 
-public class FileTypeEmbeddingProvider implements ValueProvider {
+public class VectorStoreProvider implements ValueProvider {
 
   @Override
   public Set<Value> resolve() throws ValueResolvingException {
 
-    return ValueBuilder.getValuesFor(
-            Constants.FILE_TYPE_ANY,
-            Constants.FILE_TYPE_TEXT,
-            Constants.FILE_TYPE_URL,
-            Constants.FILE_TYPE_CRAWL);
+     return ValueBuilder.getValuesFor(
+            Constants.VECTOR_STORE_PGVECTOR,
+            Constants.VECTOR_STORE_ELASTICSEARCH,
+            Constants.VECTOR_STORE_MILVUS,
+            Constants.VECTOR_STORE_CHROMA,
+            Constants.VECTOR_STORE_PINECONE,
+            Constants.VECTOR_STORE_WEAVIATE,
+            Constants.VECTOR_STORE_AI_SEARCH
+    ); // MuleChainVectorsConstants.VECTOR_STORE_NEO4J
   }
 
 }
