+++ src/main/java/org/mule/extension/mulechain/vectors/internal/storage/S3FileReader.java
@@ -10,7 +10,7 @@ import dev.langchain4j.data.document.loader.amazon.s3.AwsCredentials;
 import dev.langchain4j.data.document.DocumentParser;
 import java.util.List;
 
-import org.mule.extension.mulechain.vectors.internal.constants.MuleChainVectorsConstants;
+import org.mule.extension.mulechain.vectors.internal.constants.Constants;
 import org.mule.extension.mulechain.vectors.internal.helpers.parameters.FileTypeParameters;
 import dev.langchain4j.data.document.parser.TextDocumentParser;
 import dev.langchain4j.data.document.parser.apache.tika.ApacheTikaDocumentParser;
@@ -37,10 +37,10 @@ public class S3FileReader {
     {
         DocumentParser parser = null;
         switch (fileType.getFileType()){
-            case MuleChainVectorsConstants.FILE_TYPE_TEXT:
+            case Constants.FILE_TYPE_TEXT:
                 parser = new TextDocumentParser();
                 break;
-            case MuleChainVectorsConstants.FILE_TYPE_ANY:
+            case Constants.FILE_TYPE_ANY:
                 parser = new ApacheTikaDocumentParser();
                 break;
             default:
@@ -64,18 +64,18 @@ public class S3FileReader {
     public void readFile(String key, FileTypeParameters fileType, EmbeddingStoreIngestor ingestor) {
         DocumentParser parser = null;
         switch (fileType.getFileType()){
-            case MuleChainVectorsConstants.FILE_TYPE_TEXT:
-            case MuleChainVectorsConstants.FILE_TYPE_CRAWL:
+            case Constants.FILE_TYPE_TEXT:
+            case Constants.FILE_TYPE_CRAWL:
                 parser = new TextDocumentParser();
                 break;
-            case MuleChainVectorsConstants.FILE_TYPE_ANY:
+            case Constants.FILE_TYPE_ANY:
                 parser = new ApacheTikaDocumentParser();
                 break;
             default:
                 throw new IllegalArgumentException("Unsupported File Type: " + fileType.getFileType());
         }
         Document document = loader.loadDocument(bucketName, key, parser);
-        if (fileType.getFileType().equals(MuleChainVectorsConstants.FILE_TYPE_CRAWL)){
+        if (fileType.getFileType().equals(Constants.FILE_TYPE_CRAWL)){
             addMetadata(document);
         }
         ingestor.ingest(document);
@@ -87,9 +87,9 @@ public class S3FileReader {
             String content = jsonNode.path("content").asText();
             String source_url = jsonNode.path("url").asText();
             String title = jsonNode.path("title").asText();
-            document.metadata().add(MuleChainVectorsConstants.METADATA_KEY_FILE_TYPE, MuleChainVectorsConstants.FILE_TYPE_TEXT);
-            document.metadata().add(MuleChainVectorsConstants.METADATA_KEY_FILE_NAME, title);
-            document.metadata().add(MuleChainVectorsConstants.METADATA_KEY_FULL_PATH, source_url);
+            document.metadata().add(Constants.METADATA_KEY_FILE_TYPE, Constants.FILE_TYPE_TEXT);
+            document.metadata().add(Constants.METADATA_KEY_FILE_NAME, title);
+            document.metadata().add(Constants.METADATA_KEY_FULL_PATH, source_url);
             document.metadata().put("source", source_url);
             document.metadata().add("title", title);
         } catch (IOException e) {
